watchdog: Add watchdog via the hardware support
To check if the main loop has been stalled we add a watchdog to debug
builds. The watchdog will raise an interrupt after the "early warning"
timeout has passed and later reset the chip if the time-out period is
reached.

In this implementation the time-out period will never be reached since
the early warning interrupt eventually feeds the watchdog in the final
loop.

The purpose of this tool is to easily find out if the main loop has been
stalled. The main loop must run at least every 1ms.